Replace Docker-in-Docker setup with multi-stage build.
Docker-in-docker was being used as a way to produce artifacts in one
docker build and use them as part of the build of another image, using
different base images.

This "builder pattern" was pretty common in the mid-2010s, but has been
suplanted by multi-stage builds instead, which reduces a lot of the
complexity. See https://www.docker.com/blog/multi-stage-builds/ for the
announcement of that feature, and
https://blog.alexellis.io/mutli-stage-docker-builds/ for some of the
discussion at that time.