Add letsencrypt support.
The actual certificate request happens outside of the proxy, but in
order to prove our ownership of the domain we need to be able to host
ACME challenges at port 80. We expose the
`/var/www/.well-known/acme-challenge` directory, which can be mounted
from a volume.

There is a chicken-and-egg problem when trying to start the container
for the first time: nginx won't start without a certificate, and we
can't get a certificate without nginx running. We break that cycle by
generating a self-signed certificate if no other one is found. As soon
as the real certificate is obtained, it can be swapped in by writing it
and its key to `/etc/montagu/proxy/` and running
`docker exec proxy nginx reload -s` (or equivalent).

Instead of getting the Diffie-Hellman parameters alongside our
certificates, we now download a set of parameters. DH parameters do not
need to be secret nor unique. They only need to have been generated
"correctly". The parameters we use come from [RFC7919][rfc7919],
following the latest guidance from Mozilla. The directive only applies
to TLS1.2. Starting with TLS1.3, [the standard][rfc8446] requires the
parameters from RFC7919 to be used, and presumably nginx/openssl
hardcodes these.

[rfc7919]: https://www.rfc-editor.org/rfc/rfc7919
[rfc8446]: https://www.rfc-editor.org/rfc/rfc8446#section-4.2.8.1