Remove custom proxy-metrics code.
The only reason we had a custom implementation for proxy-metrics was
because constellation used to start containers without any networking,
and would only attach to the network as a second step, which broke the
proxy's expectation. This issue was solved in
reside-ic/constellation#23.

Thanks to this we can simplify our setup and use the standard
declarative approach even for the proxy-metrics container.